IT Department Chatbot tested with JUnit, Mockito, PowerMock,JaCoCo
This was part of my coursework in software testing we had to test at
least 10 features of  a software using a comprehensive set of software
testing tools. I used JUnit, Mockito, PowerMock and JaCoco as my main
tools for testing. Results were positive with a 87% code coverage. All
test cases function as planned.